Problem: queue does not belong to the "stack" module
Solution: extract the "queue" module
